+++ changelog.md
@@ -1,1 +1,1 @@
-- Reworked configs
+- Added a config for mobs that can spawn in the mining dimension